[WEEX-415][Android] fix waterfall layout scroll to top.

+++ android/sdk/src/main/java/com/taobao/weex/ui/component/list/WXListComponent.java
@@ -180,7 +180,7 @@ public class WXListComponent extends BasicListComponent<BounceRecyclerView> {
   }
 
   @WXComponentProp(name = Constants.Name.COLUMN_WIDTH)
-  public void setColumnWidth(int columnWidth)  {
+  public void setColumnWidth(float columnWidth)  {
     if(columnWidth != mColumnWidth){
       markComponentUsable();
       updateRecyclerAttr();
